Move tests files
