+++ tests/extensions/test_merger.py
