+++ tests/extensions/test_stages.py
